Added 2 basic routers
