--- app.ts
@@ -1,11 +1,0 @@
-const express = require("express");
-const bodyParser = require("body-parser");
-
-const app = express();
-
-app.use(bodyParser.json());
-app.use(bodyParser.urlencoded({extended: false}));
-
-app.listen(3000, () => {
-    console.log("Server listening on port 3000...");
-})
